--added a simple testing function for testing inputs from a ImGui button.
--hitting the left control key while focused on the GL window now "frees" the mouse input so that it will interact with ImGui elements.

+++ OrganicGLWinLib/WorldLightContainer.h
@@ -20,7 +20,7 @@ class WorldLightContainer
 		}
 		int getCombinedWorldLightBytes()		// return the total amount of bytes that this container uses, in regards to the combined byte size of all the world lights.
 		{
-			return (sizeof(WorldLight) * worldLights.size());
+			return int((sizeof(WorldLight) * worldLights.size()));
 		}
 };
